style: dailysentence 스타일 수정 및 드롭다운 스타일 추가
