+++ src/assets/scss/_style.scss
@@ -9,6 +9,7 @@
 @import "@scss/assets/input";
 @import "@scss/assets/padding";
 @import "@scss/assets/icon";
+@import "@scss/assets/dropdown";
 
 // pages
 @import "@scss/style/load";
