+++ src/assets/scss/assets/_dropdown.scss
@@ -1,0 +1,46 @@
+.dropdown{
+	position: relative;
+	&_toggle{
+		width: 24px;
+		height: 24px;
+		font-size: 20px;
+		color: $default-666;
+	}
+	&_wrap{
+		width: 100px;
+		position: absolute;
+		right: 0;
+		background-color: $default-fff;
+		border-radius: $main-br;
+		border: 1px solid $default-ddd;
+		overflow: hidden;
+		z-index: 10;
+	}
+	&_cont{
+		position: relative;
+		z-index: 10;
+	}
+	&_btn{
+		width: 100%;
+		display: flex;
+		align-items: center;
+		justify-content: space-between;
+		padding: 8px 12px;
+		color: $default-666;
+		border-bottom: 1px solid $default-ddd;
+		&:last-child{
+			border-bottom: 0;
+		}
+		&:hover{
+			background-color: $default-ddd;
+		}
+	}
+	&_close{
+		position: fixed;
+		top: 0;
+		left: 0;
+		width: 100%;
+		height: 100%;
+		z-index: 1;
+	}
+}
