+++ src/assets/scss/components/daily_sentence/_view.scss
@@ -3,6 +3,7 @@
 		margin-bottom: 24px;
 		display: flex;
 		align-items: center;
+		justify-content: space-between;
 		gap: 12px;
 		&_btn{
 			width: 24px;
@@ -42,6 +43,7 @@
 		font-weight: bold;
 		strong{
 			color: $main-accent-color;
+			cursor: pointer;
 		}
 	}
 	&_mean{
@@ -84,6 +86,18 @@
 			border: 1px solid $main-accent-color;
 			line-height: 1.3;
 			cursor: pointer;
+			&.on{
+				position: relative;
+				border-color: transparent;
+				&::before{
+					content: "";
+					position: absolute;
+					width: 100%;
+					height: 100%;
+					border: 3px solid $main-accent-color;
+					border-radius: $sub-br;
+				}
+			}
 			p{
 				&:last-child{
 					opacity: 0.8;
