+++ src/assets/scss/style/daily_sentence.scss
@@ -147,6 +147,7 @@
 			align-items: center;
 			border-radius: $sub-br;
 			background-color: $main-bg-color;
+			cursor: pointer;
 			&::before{
 				content: "";
 				display: block;
